// src/app/modelos/tarjeta-electronica.interface.ts
export interface TarjetaElectronica {
    idTarjetaElectronica?: number;
    cveNumeroConvenio?: string;
    nomEmpresa?: string;
    fecIniConvenio?: string;
    fecFinConvenio?: string;
    impMensual?: number;
    canLitrosLimiteMes?: number;
    idOoad?: number;
    numFolioInicial?: string;
    numFolioFinal?: string;
    canKmsRecorridos?: number;
    desEstatusTarjeta?: string;
    cveMatricula?: string;
}